Fix block producer factory registration
